+++ src/Tsufeki/Tenkawa/Server/Utils/Template.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Tsufeki\Tenkawa\Server\Utils;
+
+interface Template
+{
+    /**
+     * @param array<string,string> $variables
+     *
+     * @resolve string
+     */
+    public function render(array $variables): \Generator;
+}